# fix UBSAN's invalid-null-argument error in zstd_decompress.c