This was exposed when running another project's tests that use zstd with UBSAN instrumentation.