@terrelln 2020-01-06
Thanks for the patch! Before we merge, I would like to add a test that exposes it, since we're missing this coverage. It looks like it is triggered by a `NULL` input/output buffer in the streaming decompression API.