@terrelln 2020-01-07
> Out of curiosity, what project exposed this?

Our tests in `folly/compression/test:compression_test` when compiled with a new build mode with more UBSAN options enabled.