@terrelln 2020-01-09
The test in PR #1945 catches this bug, merging now.