@Cyan4973 2020-01-09
2 points :
- the fix in `folly` is different : the branch is `if (src != NULL)`,
which might provide better latency, since `length` must be calculated first.
- It would have been nice to measure potential speed difference introduced by this patch. I don't anticipate any, just it would be nice to measure it.